split out core interfaces into separate assembly and refactored projects to only reference the interface project and not the full core project itself
