+++ DS4Windows.Client.Core.Interfaces/DependencyInjection/IServiceRegistrar.cs
