+++ DS4Windows.Client.Core.Interfaces/View/IView.cs
