+++ DS4Windows.Client.Core/DS4Windows.Client.Core.csproj
@@ -28,4 +28,8 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\DS4Windows.Client.Core.Interfaces\DS4Windows.Client.Core.Interfaces.csproj" />
+  </ItemGroup>
+
 </Project>
