+++ DS4Windows.Client.Core/ViewModel/ViewModel.cs
@@ -19,18 +19,4 @@ namespace DS4Windows.Client.Core.ViewModel
         }
 
     }
-
-    public interface IViewModel<TViewModel> : IViewModel, INotifyPropertyChanged
-        where TViewModel : IViewModel<TViewModel>
-    {
-    }
-
-    public interface IViewModel
-    {
-        void AddView(IView view);
-
-        List<IView> Views { get; }
-
-        object? MainView { get; }
-    }
 }
