+++ DS4Windows.Shared.Common/DS4Windows.Shared.Common.csproj
@@ -39,6 +39,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\DS4Windows.Client.Core\DS4Windows.Client.Core.csproj" />
+    <ProjectReference Include="..\DS4Windows.Client.Core.Interfaces\DS4Windows.Client.Core.Interfaces.csproj" />
   </ItemGroup>
 </Project>
